Create sorted_list_active_choice
This code is intended for use in Active Choice parameters in the jenkins pipeline. It reads data from the inventory file located on disk on the jenkins server, selects group names and server names. Then sorts the lists and merges them into the resulting list. Convenient for pipelines that require host names or names of groups that combine them as a parameter.